+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Added colors to the `+` on commands with hidden subcommands in the tree graph
+- `-` is not an option anymore (again)
 
 ## v2.0.3 2024-07-14