+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 
 - Updated error message from `Error: unexpected arg` to `Error: unexpected argument`
 - Added v to versions in Changelog
-- Fixed option argumunt rendering in usage error
+- Fixed option argument rendering in usage error
 
 ## v2.0.1 2024-07-07